requirements.txt
pandas==2.0.3
SQLAlchemy==2.0.36
PyYAML==6.0.2
python-pptx==1.0.2
reportlab==4.2.5
matplotlib==3.9.2
requests==2.32.3
numpy==1.26.4